+++ src/js/control-bar/skip-buttons/skip-forward.js
@@ -17,6 +17,7 @@ class SkipForward extends Button {
     this.skipTime = this.getSkipForwardTime();
 
     if (this.skipTime && this.validOptions.includes(this.skipTime)) {
+      this.setIcon(`forward-${this.skipTime}`);
       this.controlText(this.localize('Skip forward {1} seconds', [this.skipTime]));
       this.show();
     } else {
@@ -46,6 +47,10 @@ class SkipForward extends Button {
    *        to be called
    */
   handleClick(event) {
+    if (isNaN(this.player_.duration())) {
+      return;
+    }
+
     const currentVideoTime = this.player_.currentTime();
     const liveTracker = this.player_.liveTracker;
     const duration = (liveTracker && liveTracker.isLive()) ? liveTracker.seekableEnd() : this.player_.duration();