+++ src/js/control-bar/text-track-controls/descriptions-button.js
@@ -27,6 +27,8 @@ class DescriptionsButton extends TextTrackButton {
   constructor(player, options, ready) {
     super(player, options, ready);
 
+    this.setIcon('audio-description');
+
     const tracks = player.textTracks();
     const changeHandler = Fn.bind_(this, this.handleTracksChange);